Adds a new docker function to run commands
Docker creates a self contained dependencies box, and sometimes there is
the need of make recipes to run where the dependencies are installed.
For this purpose, we can run a shell inside a container and run the make
reciped from there.
It's also possible to run the command in a non interactive way, but I
would be careful to not create recipes that rely on Docker, because this
will remove the power of running it anywhere.