Improves dotenv parsing and generation
When importing a .env with quotes to docker it would keep the quotes
which was messing up with python sys.path.

Pytho's sys.path recognises if the env var PYTHONPATH starts with a
leading slash / and consider it a full path or relative otherwise. The
problem resides when this path contains a quote.

Quotes are supposed to allow the description of a dotenv file that can
be used by `source` aka `.`. It is not supposed to keep the quotes in
the environment. That is a flaw in Docker's --env-file flag, while
make's `include` works as expected.
Now we have a function that emulates make's incluce but can be used in
recipes.

It's nice to use `env` to print the environment direct to a file. The
issue is that this file will not contain quotes, hence not compatible
with `source`.

The new .env recipe now tries to change the file as minimum as possible,
but also ensures it will contain quotes to be compatible with `source`.

Having a .env compatible with `source` makes things easier to create a
shell environment with our environment variables exported.
For that we use a Target-specific Variable called DOTENV_SHELL.

The swap_file function was a collaretal effect of having to write to the
same file you are piping, which is a common issue now sorted.