Document dependencies
The coreutils dependency is needed for some of the recipes. I had an
issue with a system that didn't had `join` installed.

In addition I'm exposing all other dependencies. We need to keep track
of all dependecies added. The cool thing is that you might not need to
install any extra dependency until you use it.