Hotfix not respecting the environment values
Make variables have precedence over environment, and the situation was
worse if a Makefile decided to include the `.env` file.

Including the `.env` file is cool because you can create recipes based
on your environment file, but the drawback is that it will override all
environments variables.

This recipe needs to dump the environment and in order to do that I had
to save the variables prior any include or most variables are defined.