Adds a new function for converting env to ini
A envfile is almost similar to a ini file. The difference lies on when
your values have spaces.
If you add quotes to a envfile they will be ignored when exported, while
quotes in a ini file will be considered part of the value.

This new function is to support ini files used by Docker and Kubernetes.

In addition, to avoid errors in the terminal envfile_to_env is now
ignoring comments and empty variables.